+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.2.1
+- FEATURE on -g to run commands on entire group (by @nertwork)
+- FEATURE on -r to thread commands run on entire group (by @nertwork)
+
 # 0.2.0
 - FEATURE on -x disable ssh exec mode (by @nickhilliard)